feat: separate app on modules
