+++ src/CLI.js
@@ -1,10 +1,13 @@
-import crypto from 'crypto';
 import readline from 'readline';
-import os, { EOL } from 'os';
-import fsPromises from 'fs/promises';
-import fs from 'fs';
+import os from 'os';
 import path from 'path';
-import { createGzip, createGunzip } from 'zlib';
+import { checkPathAccess } from './utils.js';
+import {
+  list, readFile, renameFile, copyFile, removeFile, createFile,
+} from './basicFileUtils.js';
+import { getConstant } from './constantsUtils.js';
+import { getFileHash } from './hashUtils.js';
+import { compressFile, decompressFile } from './compressUtils.js';
 
 const USER_GREETING = 'Welcome to the File Manager, <username>!\n';
 const DEFAULT_USERNAME = 'Username';
@@ -56,95 +59,6 @@ export class CLI {
     this.readlineInterface.prompt();
   }
 
-  async list(listDirPath = '') {
-    const dirPath = path.join(listDirPath);
-    return fsPromises.readdir(dirPath);
-  }
-
-  async checkPathAccess(pathToCheck, constant = fsPromises.constants.F_OK) {
-    try {
-      await fsPromises.access(pathToCheck, constant);
-      return { ok: true, err: '' };
-    } catch (err) {
-      return { ok: false, err };
-    }
-  }
-
-  async readFile(pathToFile) {
-    return (await fsPromises.readFile(pathToFile)).toString();
-  }
-
-  async createFile(pathToFile) {
-    const isAvailable = await this.checkPathAccess(path.dirname(pathToFile), fsPromises.F_OK);
-    if (isAvailable.ok) {
-      return fsPromises.writeFile(pathToFile, '', { encoding: 'utf-8', flag: 'w' });
-    }
-    throw isAvailable.err;
-  }
-
-  async renameFile(srcFile, distFile) {
-    const isAvailable = await this.checkPathAccess(srcFile, fsPromises.F_OK);
-    if (isAvailable.ok) {
-      return fsPromises.rename(srcFile, distFile);
-    }
-    throw isAvailable.err;
-  }
-
-  async copyFile(srcFile, distFile) {
-    const isAvailable = await this.checkPathAccess(srcFile, fsPromises.R_OK);
-    if (isAvailable.ok) {
-      await this.createFile(distFile);
-
-      const readableStream = fs.createReadStream(srcFile);
-      const fileStream = fs.createWriteStream(distFile, { flags: 'w' });
-
-      return readableStream.pipe(fileStream);
-    }
-    throw isAvailable.err;
-  }
-
-  async removeFile(pathToFile) {
-    await fsPromises.access(pathToFile, fsPromises.F_OK);
-
-    return fsPromises.unlink(pathToFile);
-  }
-
-  async getFileHash(pathToFile) {
-    const isAvailable = await this.checkPathAccess(pathToFile, fsPromises.R_OK);
-    if (isAvailable.ok) {
-      return new Promise((resolve, reject) => {
-        const hash = crypto.createHash('sha256');
-        const fileStream = fs.createReadStream(pathToFile);
-        fileStream.on('error', (err) => reject(err));
-        fileStream.on('data', (chunk) => hash.update(chunk));
-        fileStream.on('end', () => resolve(hash.digest('hex')));
-      });
-    }
-    throw isAvailable.err;
-  }
-
-  async compressFile(srcFilePath, distFilePath) {
-    const isAvailable = await this.checkPathAccess(srcFilePath, fsPromises.R_OK);
-    if (isAvailable.ok) {
-      const compressOutputStream = fs.createWriteStream(distFilePath);
-      const inputFileStream = fs.createReadStream(srcFilePath);
-
-      return inputFileStream.pipe(createGzip()).pipe(compressOutputStream);
-    }
-    throw isAvailable.err;
-  }
-
-  async decompressFile(srcFilePath, distFilePath) {
-    const isAvailable = await this.checkPathAccess(srcFilePath, fsPromises.R_OK);
-    if (isAvailable.ok) {
-      const decompressOutputStream = fs.createWriteStream(distFilePath);
-      const compressedInputFileStream = fs.createReadStream(srcFilePath);
-
-      return compressedInputFileStream.pipe(createGunzip()).pipe(decompressOutputStream);
-    }
-    throw isAvailable.err;
-  }
-
   async handleInput(string) {
     if (string === '.exit') {
       this.handleExit();
@@ -153,7 +67,7 @@ export class CLI {
 
     if (string === 'ls') {
       try {
-        const files = (await this.list(this.currentPath)).join('\n');
+        const files = (await list(this.currentPath)).join('\n');
         this.print(`${files}\n`);
       } catch (err) {
         this.print(`${DEFAULT_ERROR_TEXT}: ${err}\n`);
@@ -163,7 +77,7 @@ export class CLI {
 
     if (string === 'up') {
       const newPath = path.dirname(this.currentPath);
-      const pathAvailable = (await this.checkPathAccess(newPath));
+      const pathAvailable = (await checkPathAccess(newPath));
       if (pathAvailable.ok) {
         this.currentPath = newPath;
       }
@@ -173,7 +87,7 @@ export class CLI {
 
     if (string.indexOf('cd') === 0) {
       const newPath = path.resolve(this.currentPath, string.replace('cd ', ''));
-      const pathAvailable = (await this.checkPathAccess(newPath));
+      const pathAvailable = (await checkPathAccess(newPath));
       if (pathAvailable.ok) {
         this.currentPath = newPath;
         this.print('');
@@ -186,10 +100,10 @@ export class CLI {
 
     if (string.indexOf('cat') === 0) {
       const newPath = path.resolve(this.currentPath, string.replace('cat ', ''));
-      const pathAvailable = (await this.checkPathAccess(newPath));
+      const pathAvailable = (await checkPathAccess(newPath));
       if (pathAvailable.ok) {
         try {
-          const fileContents = await await this.readFile(newPath);
+          const fileContents = await await readFile(newPath);
           this.print(fileContents);
         } catch (err) {
           this.print(`${DEFAULT_ERROR_TEXT}: ${err}\n`);
@@ -205,7 +119,7 @@ export class CLI {
       const newPath = path.resolve(this.currentPath, string.replace('add ', ''));
 
       try {
-        await this.createFile(newPath);
+        await createFile(newPath);
 
         this.print(`File "${string.replace('add ', '')}" has been created\n`);
       } catch (err) {
@@ -219,7 +133,7 @@ export class CLI {
       const newPath = path.resolve(this.currentPath, string.replace('rm ', ''));
 
       try {
-        await this.removeFile(newPath);
+        await removeFile(newPath);
 
         this.print(`File "${string.replace('rm ', '')}" has been removed\n`);
       } catch (err) {
@@ -235,7 +149,7 @@ export class CLI {
         const srcFilePath = path.resolve(this.currentPath, args[0]);
         const distFilePath = path.resolve(this.currentPath, args[1]);
         try {
-          await this.renameFile(srcFilePath, distFilePath);
+          await renameFile(srcFilePath, distFilePath);
 
           this.print(`File "${args[0]}" has been renamed\n`);
         } catch (err) {
@@ -254,7 +168,7 @@ export class CLI {
         const srcFilePath = path.resolve(this.currentPath, args[0]);
         const distFilePath = path.resolve(this.currentPath, args[1]);
         try {
-          await this.copyFile(srcFilePath, distFilePath);
+          await copyFile(srcFilePath, distFilePath);
 
           this.print(`File "${args[0]}" has been copied\n`);
         } catch (err) {
@@ -273,8 +187,8 @@ export class CLI {
         const srcFilePath = path.resolve(this.currentPath, args[0]);
         const distFilePath = path.resolve(this.currentPath, args[1]);
         try {
-          await this.copyFile(srcFilePath, distFilePath);
-          await this.removeFile(srcFilePath);
+          await copyFile(srcFilePath, distFilePath);
+          await removeFile(srcFilePath);
 
           this.print(`File "${args[0]}" has been moved\n`);
         } catch (err) {
@@ -289,32 +203,12 @@ export class CLI {
 
     if (string.indexOf('os') === 0) {
       const command = string.replace('os --', '');
-      let output = '';
-      switch (command) {
-        case 'EOL': {
-          output = EOL;
-          break;
-        }
-        case 'cpus': {
-          output = `${os.cpus().length} cores:\n${os.cpus().map((el) => el.model).join('\n')}`;
-          break;
-        }
-        case 'homedir': {
-          output = os.homedir();
-          break;
-        }
-        case 'username': {
-          output = os.userInfo().username;
-          break;
-        }
-        case 'architecture': {
-          output = os.arch();
-          break;
-        }
-        default: {
-          this.print(`Unknown argument: ${command}`);
-        }
+      const output = getConstant(command);
+      if (!output) {
+        this.print(`Unknown argument: ${command}`);
+        return;
       }
+
       this.print(`${output}\n`);
       return;
     }
@@ -323,7 +217,7 @@ export class CLI {
       const newPath = path.resolve(this.currentPath, string.replace('hash ', ''));
 
       try {
-        const hash = await this.getFileHash(newPath);
+        const hash = await getFileHash(newPath);
 
         this.print(`File hash is:\n${hash}\n`);
       } catch (err) {
@@ -339,7 +233,7 @@ export class CLI {
         const srcFilePath = path.resolve(this.currentPath, args[0]);
         const distFilePath = path.resolve(this.currentPath, args[1]);
         try {
-          await this.compressFile(srcFilePath, distFilePath);
+          await compressFile(srcFilePath, distFilePath);
 
           this.print(`File "${args[0]}" has been compressed\n`);
         } catch (err) {
@@ -358,7 +252,7 @@ export class CLI {
         const srcFilePath = path.resolve(this.currentPath, args[0]);
         const distFilePath = path.resolve(this.currentPath, args[1]);
         try {
-          await this.decompressFile(srcFilePath, distFilePath);
+          await decompressFile(srcFilePath, distFilePath);
 
           this.print(`File "${args[0]}" has been decompress\n`);
         } catch (err) {
